add definitions for LASER_IN_USE_PIN and SAFETY_CUTOFF_SIGNAL_PIN
These two pins are currently not functional, just definitions for now.

Laser in use pin - signal from laser to advise openchill that the laser is currently running a job.  On the K40 using Nano M2 board, this wire will connect to EX+, and a ground is tied from OpenChill to the Nano M2 board.

SAFETY_CUTOFF_SIGNAL_PIN - this signal wire is to tell the laser that OpenChill needs it to pause/cease operation as OpenChill cannot achieve the desired chiller temperature, or needs a break to cool down a bit.  On the K40 with standard PSUs this refers to the Power supply pin P+, with P- being tied to ground on OpenChill